Topic Modeling
The notebook is focused on topic modeling, beginning with the essential setup for data analysis and visualization. It imports standard libraries like NumPy and pandas for data manipulation, matplotlib and seaborn for plotting and visual aesthetics, and scipy for advanced mathematical operations. The inclusion of libraries such as swifter suggests an emphasis on performance optimization, allowing for faster application of functions on pandas data structures. The setup indicates a preparation for handling and visualizing large datasets, likely to explore patterns, trends, and distributions within textual data as part of the topic modeling process. This foundation suggests a workflow that not only prioritizes efficient data processing but also detailed and insightful data visualization to support the identification and analysis of topics within a corpus